Very nice code that doesn't work; At least few misconceptions fixed anyway.

+++ include/stdint.h
@@ -16,7 +16,7 @@
 	typedef long 		int32_t;
 	typedef long long 	int64_t;
 	typedef unsigned char 	uint8_t;
-	typedef unsigned int 	uint16_t;
+	typedef unsigned short 	uint16_t;
 	typedef unsigned long 	uint32_t;
 	typedef unsigned long long uint64_t;
 #else
